Fixed according to comments
